Move TemplateEditor
