+++ site/src/pages/TemplateVersionEditorPage/TemplateVersionEditor/FileTreeView.tsx
