+++ site/src/pages/TemplateVersionEditorPage/TemplateVersionEditor/MonacoEditor.tsx
