+++ site/src/pages/TemplateVersionEditorPage/TemplateVersionEditor/PublishTemplateVersionDialog.tsx
@@ -4,7 +4,7 @@ import { getFormHelpers } from "utils/formUtils"
 import { FormFields } from "components/Form/Form"
 import { useFormik } from "formik"
 import * as Yup from "yup"
-import { PublishVersionData } from "pages/TemplateVersionPage/TemplateVersionEditorPage/types"
+import { PublishVersionData } from "pages/TemplateVersionEditorPage/types"
 import TextField from "@mui/material/TextField"
 import { ConfirmDialog } from "components/Dialogs/ConfirmDialog/ConfirmDialog"
 import Checkbox from "@mui/material/Checkbox"
